+++ freshbooks/api/events.py
@@ -1,3 +1,4 @@
+from collections import namedtuple
 from decimal import Decimal
 from typing import Any, Optional
 
@@ -20,6 +21,19 @@ class EventsResource(AccountingResource):
                 self.base_url, account_id, self.accounting_path, resource_id)
         return "{}/events/account/{}/{}".format(self.base_url, account_id, self.accounting_path)
 
+    def _extract_error(self, errors: dict) -> namedtuple:
+        Error = namedtuple("Error", "message details code")
+
+        if not errors:  # pragma: no cover
+            return Error("Unknown error", None, None)
+
+        details = None
+        for detail in errors.get("details", []):
+            if detail.get("@type") == "type.googleapis.com/google.rpc.BadRequest":
+                details = detail
+
+        return Error(errors["message"], details, int(errors["code"]))
+
     def _request(self, url: str, method: str, data: Optional[dict] = None) -> Any:
         response = self._send_request(url, method, data)
 
@@ -36,8 +50,10 @@ class EventsResource(AccountingResource):
             raise FreshBooksError(status, "Failed to parse response", raw_response=response.text)
 
         if status >= 400:
-            message, code = self._extract_error(content)
-            raise FreshBooksError(status, message, error_code=code, raw_response=content)
+            error = self._extract_error(content)
+            raise FreshBooksError(
+                status, error.message, error_code=error.code, error_details=error.details, raw_response=content
+            )
 
         if "response" not in content:
             raise FreshBooksError(status, "Returned an unexpected response", raw_response=response.text)